Move new components to correct place'
