+++ src/index.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-// import App from './App';
-import App from './NewApp';
+import App from './App';
 import './index.css';
 
 ReactDOM.render(
